Integrate this feature into envguard as a new function:  analyzeEnv({   envPath?: string,   rootDir?: string })  Default: - envPath = ".env" - rootDir = process.cwd()  It should: 1. Parse .env 2. Scan codebase 3. Compare 4. Print report  Do not break existing guard() or inferEnv().
