+++ src/index.js
@@ -1,7 +1,7 @@
 import { guard } from "./guard.js";
 import { inferEnv } from "./infer.js";
 import { coerce, inferType, isBoolean, isNumeric, parseEnv } from "./utils.js";
-import { scanUsedVars, analyzeEnv } from "./scanner.js";
+import { scanUsedVars, compareEnvVars, analyzeEnv } from "./scanner.js";
 
 export { 
   guard, 
@@ -12,5 +12,6 @@ export {
   isNumeric, 
   parseEnv,
   scanUsedVars,
+  compareEnvVars,
   analyzeEnv
 };
